# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Standout Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get assessment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assessment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{id}/submit:
    post:
      operationId: submitAssessment
      tags: [assessments]
      summary: Submit assessment answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentSubmitBody"
      responses:
        "200":
          description: Assessment results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"
...
  /scenarios:
    get:
      operationId: listScenarios
      tags: [scenarios]
      summary: List scenarios
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
  /scenarios/{id}:
    get:
      operationId: getScenario
      tags: [scenarios]
      summary: Get scenario by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scenario detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
...
  /systems:
    get:
      operationId: listSystems
      tags: [systems]
      summary: List management systems/playbooks
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/System"
...
  /admin/scenarios:
    get:
      operationId: adminListScenarios
      tags: [admin]
      summary: Admin list all scenarios
      responses:
        "200":
          description: All scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
    post:
      operationId: adminCreateScenario
      tags: [admin]
      summary: Admin create scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioBody"
      responses:
        "201":
          description: Created scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
  /admin/scenarios/{id}:
    put:
      operationId: adminUpdateScenario
      tags: [admin]
      summary: Admin update scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioBody"
      responses:
        "200":
          description: Updated scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
    delete:
      operationId: adminDeleteScenario
      tags: [admin]
      summary: Admin delete scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/systems:
    get:
      operationId: adminListSystems
      tags: [admin]
      summary: Admin list all systems
      responses:
        "200":
          description: All systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/System"
    post:
      operationId: adminCreateSystem
      tags: [admin]
      summary: Admin create system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemBody"
      responses:
        "201":
          description: Created system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        tier:
          type: string
          enum: [free, core, premium]
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
      required: [id, email, role, tier, onboardingCompleted, createdAt]
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required: [email, password, name]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
...
    Assessment:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        estimatedMinutes:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentQuestion"
        isCompleted:
          type: boolean
        completedAt:
          type: string
          nullable: true
        score:
          type: number
          nullable: true
      required: [id, title, description, category, estimatedMinutes, questions, isCompleted]
    AssessmentQuestion:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentOption"
        type:
          type: string
          enum: [single, scale, multi]
      required: [id, text, options, type]
    AssessmentOption:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        value:
          type: integer
      required: [id, text, value]
    AssessmentSubmitBody:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              selectedOptionId:
                type: string
              value:
                type: integer
            required: [questionId]
      required: [answers]
    AssessmentResult:
      type: object
      properties:
        id:
          type: string
        assessmentId:
          type: string
        assessmentTitle:
          type: string
        score:
          type: number
        scoreLabel:
          type: string
        completedAt:
          type: string
        strengths:
          type: array
          items:
            type: string
        blindSpots:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        breakdown:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              score:
                type: number
              label:
                type: string
            required: [category, score, label]
      required: [id, assessmentId, assessmentTitle, score, scoreLabel, completedAt, strengths, blindSpots, recommendations, breakdown]
    Scenario:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        situation:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        estimatedMinutes:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/ScenarioOption"
        tags:
          type: array
          items:
            type: string
        isCompleted:
          type: boolean
        userScore:
          type: integer
          nullable: true
        isPremium:
          type: boolean
      required: [id, title, description, situation, category, difficulty, estimatedMinutes, options, tags, isCompleted, isPremium]
    ScenarioOption:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        quality:
          type: string
          enum: [strong, acceptable, weak, risky]
        explanation:
          type: string
        brandImpact:
          type: string
        score:
          type: integer
      required: [id, text, quality, explanation, brandImpact, score]
...
    ScenarioBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        situation:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        estimatedMinutes:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/ScenarioOption"
        tags:
          type: array
          items:
            type: string
        isPremium:
          type: boolean
      required: [title, description, situation, category, difficulty, estimatedMinutes, options, tags, isPremium]
    System:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/SystemStep"
        keyPrinciples:
          type: array
          items:
            type: string
        scripts:
          type: array
          items:
            $ref: "#/components/schemas/SystemScript"
        estimatedMinutes:
          type: integer
        isPremium:
          type: boolean
      required: [id, title, description, category, steps, keyPrinciples, scripts, estimatedMinutes, isPremium]
    SystemStep:
      type: object
      properties:
        order:
          type: integer
        title:
          type: string
        description:
          type: string
      required: [order, title, description]
    SystemScript:
      type: object
      properties:
        tone:
          type: string
        text:
          type: string
      required: [tone, text]
    SystemBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/SystemStep"
        keyPrinciples:
          type: array
          items:
            type: string
        scripts:
          type: array
          items:
            $ref: "#/components/schemas/SystemScript"
        estimatedMinutes:
          type: integer
        isPremium:
          type: boolean
      required: [title, description, category, steps, keyPrinciples, scripts, estimatedMinutes, isPremium]
...