# lib/api-zod/src/generated/types/systemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Standout Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { SystemScript } from "./systemScript";
import type { SystemStep } from "./systemStep";

export interface SystemBody {
  title: string;
  description: string;
  category: string;
  steps: SystemStep[];
  keyPrinciples: string[];
  scripts: SystemScript[];
  estimatedMinutes: number;
  isPremium: boolean;
}