# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Standout Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserTier } from "./userTier";

export interface User {
  id: string;
  email: string;
  name?: string;
  role: UserRole;
  tier: UserTier;
  onboardingCompleted: boolean;
  createdAt: string;
}